Update README.md
#3 Functions with No Parameters

In the main function, rewrite the code to call the functions in the correct order so that the output makes sense chronologically.
Call the correct function so that the program says goodbye to the user.
Define a new function that tells the user a joke (you decide on the function name and the joke). Call it in a sensible place in the program (The user doesn't need to input a response).